requirements.txt
connexion[swagger-ui,uvicorn]
mariadb
argon2-cffi
